Move audio_ast/mod.rs code to (new file) data.rs
src/function/audio_ast/mod.rs contains a number of fns that better belong
in a module.  Move the code to a new file data.rs so mod.rs is only used
for defining the module.

I used git mv mod.rs data.rs to try to retain the history.
